Add permuter composition layer (Phase 3) and complete evolution
Complete all three phases of the permuter evolution:
- Phase 1: SourceEditor + ast_queries primitives
- Phase 2: Migrate all 12 patterns to use primitives
- Phase 3: Composition layer (--compose) for chaining pattern pairs

New modules: composer.py, editor.py, ast_queries.py, diagnosis.py
New patterns: argument_swap, branch_polarity, commutative_swap,
  comparison_equivalence, comparison_flip, declaration_reorder,
  empty_size_swap, fma_reorder, ternary_swap
New CLI flags: --compose, --no-guided
Test suite: 82 tests passing

Also includes: guided permuter docs, unicorn runner typed fixtures
and prober, decomp improvements from permuter batch runs, Ghidra
MCP client updates, and struct_db tooling.

+++ src/lazer/meta_ham/FitnessCalorieSort.cpp
@@ -97,7 +97,7 @@ void FitnessCalorieSort::BuildTree() {
 
     // Populate nodes from TheFitnessCalorieSortMgr calories
     std::vector<int> &calories = TheFitnessCalorieSortMgr->GetUnk78();
-    for (int i = 0; i < (int)calories.size(); i++) {
+    for (int i = 0; (unsigned int)i < (int)calories.size(); i++) {
         NavListItemNode *node = NewItemNode((void *)&calories[i]);
         nodes.push_back(node);
     }
